Split twitter API interaction to separate module
